Move handlers from 'main' package; fix GET improve method - remove redunant method validation
